ai-agents/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
langchain==0.1.0
langchain-google-genai==0.0.6
pydantic==2.5.0
httpx==0.25.2
python-dotenv==1.0.0
pandas==2.1.3
psycopg2-binary==2.9.9
asyncpg==0.29.0
